Back-end/requirements.txt
pyjwt==2.10.1
django-filter==24.3

